move BoolCell into the components directory
