+++ client/src/components/BoolCell.jsx
